Move Slack notifications into separated class
